Rename `quickStart` to `quick-start`
